requirements.txt
spotipy
twitchio
python-dotenv